% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFormulations.R
\name{A_rc_over_sd_vs_b_times_sd}
\alias{A_rc_over_sd_vs_b_times_sd}
\title{Jacobian of the risk concentration vector wrt the portfolio weights for the
formulation "rc-over-sd-vs-b-times-sd"}
\usage{
A_rc_over_sd_vs_b_times_sd(w, Sigma, N, r, b)
}
\arguments{
\item{w}{portfolio weights}

\item{Sigma}{covariance or correlation matrix}

\item{N}{number of stocks (length of w)}

\item{r}{the quantity w * (Sigma %*% w)}

\item{b}{budget vector}
}
\value{
Jg the jacobian of the risk concentration vector wrt to w
}
\description{
Jacobian of the risk concentration vector wrt the portfolio weights for the
formulation "rc-over-sd-vs-b-times-sd"
}
