% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFormulations.R
\name{A_rc_over_var_vs_b}
\alias{A_rc_over_var_vs_b}
\title{Jacobian of the risk concentration vector wrt the portfolio weights for the
formulation "rc-over-var-vs-b"}
\usage{
A_rc_over_var_vs_b(w, Sigma, N, r)
}
\arguments{
\item{w}{portfolio weights}

\item{Sigma}{covariance or correlation matrix}

\item{N}{number of stocks (length of w)}
}
\value{
Jg the jacobian of the risk concentration vector wrt to w
}
\description{
Jacobian of the risk concentration vector wrt the portfolio weights for the
formulation "rc-over-var-vs-b"
}
