% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computationAmatrix.R
\name{compute_A_dan_version}
\alias{compute_A_dan_version}
\title{function used for sanity-check / unit testing}
\usage{
compute_A_dan_version(w, Sigma, N)
}
\description{
function used for sanity-check / unit testing
}
