% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successiveConvexApprox.R
\name{compute_A_single_index_R}
\alias{compute_A_single_index_R}
\title{Computes the A matrix for the single-index formulation}
\usage{
compute_A_single_index_R(w, Sigma, N)
}
\description{
Computes the A matrix for the single-index formulation
}
