% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFormulations.R
\name{g_rc_double_index}
\alias{g_rc_double_index}
\title{Risk concentration vector for the formulation rc-double-index}
\usage{
g_rc_double_index(w, Sigma, N, r)
}
\arguments{
\item{w}{portfolio weights}

\item{Sigma}{covariance or correlation matrix}

\item{N}{number of stocks (length of w)}

\item{r}{the quantity w * (Sigma %*% w)}
}
\value{
g risk concentration vector
}
\description{
Risk concentration vector for the formulation rc-double-index
}
