% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFormulations.R
\name{g_rc_over_sd_vs_b_times_sd}
\alias{g_rc_over_sd_vs_b_times_sd}
\title{Risk vector for the formulation "rc-over-sd-vs-b-times-sd"}
\usage{
g_rc_over_sd_vs_b_times_sd(w, Sigma, r, b)
}
\arguments{
\item{w}{portfolio weights}

\item{Sigma}{covariance or correlation matrix}

\item{r}{the quantity w * (Sigma %*% w)}

\item{b}{budget vector}

\item{N}{number of stocks (length of w)}
}
\value{
g risk concentration vector
}
\description{
Risk vector for the formulation "rc-over-sd-vs-b-times-sd"
}
