% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFormulations.R
\name{R_grad_rc_double_index}
\alias{R_grad_rc_double_index}
\title{Gradient of the objective function wrt the portfolio weights for the
formulation "rc-double-index"}
\usage{
R_grad_rc_double_index(w, Sigma, N)
}
\arguments{
\item{w}{portfolio weights}

\item{Sigma}{covariance or correlation matrix}

\item{N}{number of stocks (length of w)}
}
\value{
grad_R the gradient of the obejctive function wrt to w
}
\description{
Gradient of the objective function wrt the portfolio weights for the
formulation "rc-double-index"
}
