% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFormulations.R
\name{R_grad_rc_over_sd_vs_b_times_sd}
\alias{R_grad_rc_over_sd_vs_b_times_sd}
\title{Gradient of the objective function wrt the portfolio weights for the
formulation "rc-over-sd-vs-b-times-sd"}
\usage{
R_grad_rc_over_sd_vs_b_times_sd(w, Sigma, N, b)
}
\arguments{
\item{w}{portfolio weights}

\item{Sigma}{covariance or correlation matrix}

\item{N}{number of stocks (length of w)}

\item{b}{budget vector}
}
\value{
grad_R the gradient of the obejctive function wrt to w
}
\description{
Gradient of the objective function wrt the portfolio weights for the
formulation "rc-over-sd-vs-b-times-sd"
}
