% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFormulations.R
\name{R_rc_double_index}
\alias{R_rc_double_index}
\title{Objective function for the formulation "rc-double-index"}
\usage{
R_rc_double_index(w, Sigma, N)
}
\arguments{
\item{w}{portfolio weights}

\item{Sigma}{covariance or correlation matrix}

\item{N}{number of stocks (length of w)}
}
\value{
R scalar valued objective function
}
\description{
Objective function for the formulation "rc-double-index"
}
