% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFormulations.R
\name{R_rc_over_sd_vs_b_times_sd}
\alias{R_rc_over_sd_vs_b_times_sd}
\title{Objective function for the formulation "rc-over-sd-vs-b-times-sd"}
\usage{
R_rc_over_sd_vs_b_times_sd(w, Sigma, N, b)
}
\arguments{
\item{w}{portfolio weights}

\item{Sigma}{covariance or correlation matrix}

\item{N}{number of stocks (length of w)}

\item{b}{budget vector}
}
\value{
R scalar valued objective function
}
\description{
Objective function for the formulation "rc-over-sd-vs-b-times-sd"
}
