% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successiveConvexApprox.R
\name{rho_grad}
\alias{rho_grad}
\title{Gradient of the approximation for the lp-norm, 0 < p < 1, w.r.t. to x}
\usage{
rho_grad(x, p, e)
}
\description{
Gradient of the approximation for the lp-norm, 0 < p < 1, w.r.t. to x
}
