% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskParityPortfolio.R
\name{riskParityPortfolioCVX}
\alias{riskParityPortfolioCVX}
\title{Implements the risk parity portfolio using SCA and CVXR}
\usage{
riskParityPortfolioCVX(mu, Sigma, nu = 0, budget = TRUE,
  shortselling = FALSE, w0 = NA, gamma = 0.9, zeta = 1e-07,
  tau = NA, lambda = 0.5, maxiter = 500, ftol = 1e-05,
  wtol = 1e-05)
}
\description{
Implements the risk parity portfolio using SCA and CVXR
}
