% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskParityPortfolio.R
\name{riskParityPortfolioGenSolver}
\alias{riskParityPortfolioGenSolver}
\title{Implements the risk parity portfolio using a general constrained
solver from the alabama package}
\usage{
riskParityPortfolioGenSolver(Sigma, b = rep(1/nrow(Sigma), nrow(Sigma)),
  budget = TRUE, shortselling = FALSE,
  formulation = "rc-over-var vs b", method = "slsqp",
  use_gradient = TRUE, w0 = NA, theta0 = NA, maxiter = 500,
  ftol = 1e-09, wtol = 1e-06)
}
\description{
Implements the risk parity portfolio using a general constrained
solver from the alabama package
}
