% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskParityPortfolio.R
\name{riskParityPortfolioNewton}
\alias{riskParityPortfolioNewton}
\title{Fast vanilla risk parity portfolio design using the Newton method}
\usage{
riskParityPortfolioNewton(Sigma, b = rep(1/nrow(Sigma), nrow(Sigma)),
  maxiter = 50, tol = 1e-06)
}
\arguments{
\item{Sigma}{covariance or correlation matrix}

\item{b}{budget vector}

\item{maxiter}{maximum number of iterations of both damped and quadratic phases}

\item{tol}{tolerance of the stopping criteria}
}
\value{
a list containing the following elements:
\item{\code{w}}{optimal portfolio vector}
\item{\code{risk_contribution}}{the risk contribution of every asset}
}
\description{
Risk parity portfolio optimization using the Newton method
             proposed by Spinu (2013)
}
