% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskParityPortfolio.R
\name{riskParityPortfolioQP}
\alias{riskParityPortfolioQP}
\title{Implements the risk parity portfolio using SCA and CVXR}
\usage{
riskParityPortfolioQP(Sigma, w0 = NA, gamma = 0.9, zeta = 1e-07,
  tau = NA, maxiter = 500, ftol = 1e-05, wtol = 1e-05)
}
\description{
Implements the risk parity portfolio using SCA and CVXR
}
