% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseSuccessiveConvexApprox.R
\name{theta_update}
\alias{theta_update}
\title{Update step for the average RCs of the selected assets}
\usage{
theta_update(theta_k, w_k, Sigma, g, p, e, gamma)
}
\arguments{
\item{theta_k}{the avg RCs at the previous iteration}

\item{w_k}{the portfolio weights at the previous iteration}

\item{Sigma}{covariance matrix of the assets}

\item{g}{function to represent the risk controbution}

\item{p}{parameter for the l0 norm approximation}

\item{e}{parameter for the l0 norm approximation}

\item{gamma}{the learning rate}
}
\description{
Update step for the average RCs of the selected assets
}
