% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseSuccessiveConvexApprox.R
\name{w_update}
\alias{w_update}
\title{Update step for the portfolio weights}
\usage{
w_update(w_k, theta_k, nu, gamma, l1, l2, mu, Sigma, tau, p, e, type)
}
\arguments{
\item{w_k}{weights vector at the previous iteration}

\item{theta_k}{avg the portfolio weights at the previous iteration}

\item{gamma}{the learning rate}
}
\description{
Update step for the portfolio weights
}
