% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computationAmatrix.R
\name{compute_A_double_index_R}
\alias{compute_A_double_index_R}
\title{Computes the A matrix for the formulation "rc-over-var vs b"
Also for sanity-check / unit testing}
\usage{
compute_A_double_index_R(w, Sigma, N)
}
\description{
Computes the A matrix for the formulation "rc-over-var vs b"
Also for sanity-check / unit testing
}
