% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskParityPortfolio.R
\name{riskParityPortfolio}
\alias{riskParityPortfolio}
\title{Design of Risk Parity Portfolios}
\usage{
riskParityPortfolio(Sigma, b = NULL, mu = NULL, lmd_mu = 1e-04,
  lmd_var = 0, budget = TRUE, shortselling = FALSE,
  method = c("sca", "alabama", "slsqp"), formulation = NULL,
  w0 = NULL, theta0 = NULL, gamma = 0.9, zeta = 1e-07,
  tau = NULL, maxiter = 500, ftol = 1e-06, wtol = 1e-06,
  use_gradient = TRUE)
}
\arguments{
\item{Sigma}{covariance or correlation matrix}

\item{b}{budget vector, aka, risk budgeting targets}

\item{mu}{vector of expected returns}

\item{lmd_mu}{scalar the controls the importance of the expected return term}

\item{budget}{boolean indicating whether to consider sum(w) = 1 as a
constraint}

\item{shortselling}{boolean indicating whether to allow short-selling, i.e.,
w < 0}

\item{method}{which optimization method to use.}

\item{formulation}{string indicating the formulation to be used for the risk
parity optimization problem. It must be one of: "diag", "rc-double-index",
"rc-over-b-double-index", "rc-over-var vs b", "rc-over-var",
"rc-over-sd vs b-times-sd", "rc vs b-times-var", "rc vs theta", or
"rc-over-b vs theta". If formulation is NULL and no additional terms
or constraints are set, such as expected return or shortselling, then
the vanilla risk parity portfolio will be returned. If formulation is
"diag" then the analytical solution of the risk parity optimization for
for a diagonal covariance matrix will be returned.}

\item{w0}{initial value for the portfolio wieghts. Default is the optimum
portfolio weights for the case when Sigma is diagonal.}

\item{theta0}{initial value for theta. If NULL, the optimum solution for a fixed
vector of portfolio weights will be used}

\item{gamma}{learning rate}

\item{zeta}{factor used to decrease the learning rate at each iteration}

\item{tau}{regularization factor. If NULL, a meaningful value will be used}

\item{maxiter}{maximum number of iterations for the SCA loop}

\item{ftol}{convergence tolerance on the value of the objective function}

\item{wtol}{convergence tolerance on the values of the parameters}

\item{use_gradient}{(this parameter is meaningful only if method is either
"alabama" or "slsqp") if TRUE, gradients of the objective function wrt to the
parameters will be used. This is strongly recommended to achieve faster
results.}
}
\value{
a list containing possibly the following elements:
\item{\code{w}}{optimal portfolio vector}
\item{\code{theta}}{the optimal value for theta (in case that it is part of the chosen formulation}
\item{\code{obj_fun}}{the sequence of values from the objective function at each iteration}
\item{\code{elapsed_time}}{elapsed time recorded at every iteration}
\item{\code{convergence}}{flag to indicate whether or not the optimization converged.
The value `1` means it has converged, and `0` otherwise.}
\item{\code{risk_contribution}}{the risk contribution of every asset}
}
\description{
This functions uses the vanilla (Newton) risk parity portfolio
             in order to initialize either the SCA method or a general solver.
             This convenience function helps to put together the advantages
             of both methods.
}
\author{
Ze Vinicius and Daniel P. Palomar
}
