% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskParityPortfolio.R
\name{riskParityPortfolio}
\alias{riskParityPortfolio}
\title{Jointly estimates a portifolio with risk-parity strategy and asset selection}
\usage{
riskParityPortfolio(mu, Sigma, w0 = NA, risk_contrib = NA,
  gamma = 0.1, zeta = 0.01, nu = 0.5, l1 = 0.1, l2 = 4,
  tau = 0.001, type = "1", p = 0.002, e = 1e-08, maxiter = 5000,
  w_tol = 1e-04, theta_tol = 1e-04, ftol = 1e-04)
}
\arguments{
\item{mu}{mean vector of the random returns of n assets}

\item{Sigma}{positive definite covariance matrix of the random returns
of n assets}

\item{w0}{initial estimate of the portfolio weights}

\item{risk_contrib}{function that takes the portfolio weights and the
covariance matrix of the returns as inputs and outputs the risk
contribution of each asset}

\item{gamma}{learning rate for the optimization}

\item{zeta}{how much to decrease the learning rate per iteration such that
gamma <- gamma * (1 - zeta * gamma)}

\item{nu}{regularization parameter}

\item{l1}{regularization parameter}

\item{l2}{regularization parameter}

\item{tau}{regularization parameter}

\item{type}{approximation type}

\item{maxiter}{maximum number of iterations}

\item{w_tol}{convergence tolerance on the portfolio weights}

\item{theta_tol}{convergence tolerance on the average of the RCs for the
selected assets}

\item{ftol}{convergence tolerance on the objective function}
}
\description{
Jointly estimates a portifolio with risk-parity strategy and asset selection
}
