% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_parity_cvx.R
\name{riskParityPortfolioCVX}
\alias{riskParityPortfolioCVX}
\title{Implements the risk parity portfolio using SCA and CVXR
this implementation is basically used for the sake of unit
testing.}
\usage{
riskParityPortfolioCVX(Sigma, w0 = NA, budget = TRUE,
  shortselling = FALSE, gamma = 0.9, zeta = 1e-07, tau = NA,
  maxiter = 500, ftol = 1e-05, wtol = 1e-05)
}
\description{
Implements the risk parity portfolio using SCA and CVXR
this implementation is basically used for the sake of unit
testing.
}
