% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskParityPortfolio.R
\name{riskParityPortfolioDiagSigma}
\alias{riskParityPortfolioDiagSigma}
\title{Risk parity portfolio design for uncorrelated assets}
\usage{
riskParityPortfolioDiagSigma(Sigma, b = rep(1/nrow(Sigma), nrow(Sigma)))
}
\arguments{
\item{Sigma}{covariance or correlation matrix}

\item{b}{budget vector}
}
\value{
w optimal portfolio vector

risk_contribution the risk contribution of every asset
}
\description{
Risk parity portfolio optimization for the case of diagonal Sigma
that satisfies the constraints sum(w) = 1 and w >= 0.
}
\examples{
library(riskParityPortfolio)

generate synthetic covariance matrix
N <- 100
v <- rnorm(N)
Sigma <- diag(v * v)
# compute optimal risk parity portfolio
portfolio <- riskParityPortfolioDiagSigma(Sigma)

}
