% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskParityPortfolio.R
\name{riskParityPortfolioDiagSigma}
\alias{riskParityPortfolioDiagSigma}
\title{Implements the risk parity portfolio for the case of diagonal Sigma
that satisfies the constraints sum(w) = 1 and w >= 0.}
\usage{
riskParityPortfolioDiagSigma(Sigma)
}
\description{
Implements the risk parity portfolio for the case of diagonal Sigma
that satisfies the constraints sum(w) = 1 and w >= 0.
}
