% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskParityPortfolio.R
\name{riskParityPortfolioGenSolver}
\alias{riskParityPortfolioGenSolver}
\title{Implements the risk parity portfolio using a general constrained
solver from the alabama package}
\usage{
riskParityPortfolioGenSolver(Sigma, w0 = NA, budget = TRUE,
  shortselling = FALSE)
}
\description{
Implements the risk parity portfolio using a general constrained
solver from the alabama package
}
