% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskParityPortfolio.R
\name{riskParityPortfolioGenSolver}
\alias{riskParityPortfolioGenSolver}
\title{Implements the risk parity portfolio using a general constrained
solver from the alabama package}
\usage{
riskParityPortfolioGenSolver(Sigma, w0 = NA, budget = TRUE,
  shortselling = FALSE, formulation = "rc-over-var-vs-b",
  method = "slsqp", use_gradient = TRUE, maxiter = 500, ftol = 1e-09,
  wtol = 1e-06)
}
\arguments{
\item{formulation}{a string indicating the formulation to use for the risk
parity optimization problem. It must be one of c("rc-double-index",
"rc-over-var-vs-b", "rc-over-sd-vs-b-times-sd")}
}
\description{
Implements the risk parity portfolio using a general constrained
solver from the alabama package
}
