% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/successiveConvexApprox.R
\name{theta_update}
\alias{theta_update}
\title{Update step for the average RCs of the selected assets}
\usage{
theta_update(theta_k, w_k, Sigma, g, p, e, gamma)
}
\arguments{
\item{g}{function to represent the risk controbution}

\item{gamma}{the learning rate}

\item{theta}{the avg RCs at the previous iteration}

\item{w}{the portfolio weights at the previous iteration}
}
\description{
Update step for the average RCs of the selected assets
}
